requirements.txt
aiogram>=3.0.0
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
apscheduler>=3.10.0
